refactor: MyPage 하나로 통일
