+++ src/pages/MyPage/index.ts
@@ -1,0 +1,3 @@
+import MyPage from './MyPage';
+
+export { MyPage };
