--- src/pages/MyPages/MenteeMyPage/index.ts
@@ -1,3 +1,0 @@
-import MenteeMyPage from './MenteeMyPage';
-
-export { MenteeMyPage };
