--- src/pages/MyPages/MentorMyPage/MentorMyPage.tsx
@@ -1,5 +1,0 @@
-const MentorMyPage = () => {
-  return <></>;
-};
-
-export default MentorMyPage;
