--- src/pages/MyPages/MentorMyPage/index.ts
@@ -1,3 +1,0 @@
-import MentorMyPage from './MentorMyPage';
-
-export { MentorMyPage };
